feat: fixing some things
