+++ src/main/java/com/example/schoolservlet/servlets/student/SignUpServlet.java
@@ -5,6 +5,7 @@ import com.example.schoolservlet.exceptions.DataException;
 import com.example.schoolservlet.exceptions.NotFoundException;
 import com.example.schoolservlet.exceptions.ValidationException;
 import com.example.schoolservlet.models.Student;
+import com.example.schoolservlet.utils.InputNormalizer;
 import com.example.schoolservlet.utils.InputValidation;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -14,8 +15,14 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-@WebServlet(name = "SignUpServlet", value = "/pages/students/register")
+@WebServlet(name = "SignUpServlet", value = "/student/register")
 public class SignUpServlet extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+
+        request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
+    }
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
@@ -23,7 +30,8 @@ public class SignUpServlet extends HttpServlet {
 
         String name = request.getParameter("name");
         String email = request.getParameter("email");
-        String password = request.getParameter("password");
+        String password = request.getParameter("newPassword");
+        String confirmPassword = request.getParameter("confirmPassword");
         String cpf = request.getParameter("cpf");
 
         Student newStudent = new Student();
@@ -32,22 +40,31 @@ public class SignUpServlet extends HttpServlet {
         try {
             newStudent = dao.findByCpf(cpf);
         } catch (ValidationException ve) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", ve.getMessage());
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         } catch (NotFoundException nfe) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", nfe.getMessage());
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         } catch (DataException de) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", de.getMessage());
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         }
 
-        newStudent.setName(name);
-
         try {
+            InputValidation.validateStudentName(name);
+            newStudent.setName(InputNormalizer.normalizeName(name));
+
             InputValidation.validateEmail(email);
             newStudent.setEmail(email);
+
             InputValidation.validatePassword(password);
+            InputValidation.validatePassword(confirmPassword);
+
+            if (!password.equals(confirmPassword)) throw new ValidationException("Senhas informadas devem ser iguais");
+
             newStudent.setPassword(password);
         } catch (ValidationException ve) {
             request.setAttribute("error", ve.getMessage());
@@ -57,15 +74,17 @@ public class SignUpServlet extends HttpServlet {
 
         try {
             dao.enrollIn(newStudent);
-            response.sendRedirect(request.getContextPath() + "/pages/students/signup.jsp?register=success");
-            return;
+            response.sendRedirect(request.getContextPath() + "/index.jsp");
         } catch (ValidationException ve) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", ve.getMessage());
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         } catch (NotFoundException nfe) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", "Matrícula Inválida");
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         } catch (DataException de) {
+            request.setAttribute("student", newStudent);
             request.setAttribute("error", de.getMessage());
             request.getRequestDispatcher("/pages/students/signup.jsp").forward(request, response);
         }
