+++ src/main/webapp/index.jsp
@@ -45,6 +45,7 @@
     </form>
 
     <p>Esqueceu sua senha? <a href="${pageContext.request.contextPath}/auth/forgot-password/send-code">Clique aqui</a></p>
+    <p>Não tem conta? <a href="${pageContext.request.contextPath}/student/validate/cpf">Crie aqui</a></p>
 </main>
 
 </body>
